+++ packages/amplify-category-geo/src/service-stacks/mapStack.ts
@@ -1,18 +1,21 @@
 import * as cdk from '@aws-cdk/core';
 import * as iam from '@aws-cdk/aws-iam';
 import * as lambda from '@aws-cdk/aws-lambda';
-import { MapParameters } from '../service-utils/mapParams';
 import { CfnResource, Duration, Fn } from '@aws-cdk/core';
-import { AccessType } from '../service-utils/resourceParams';
-import { BaseStack, TemplateMappings } from './baseStack';
 import { Effect } from '@aws-cdk/aws-iam';
 import { Runtime } from '@aws-cdk/aws-lambda';
-import { customMapLambdaCodePath } from '../service-utils/constants';
 import * as fs from 'fs-extra';
+import { MapParameters } from '../service-utils/mapParams';
+import { AccessType } from '../service-utils/resourceParams';
+import { BaseStack, TemplateMappings } from './baseStack';
+import { customMapLambdaCodePath } from '../service-utils/constants';
 
 type MapStackProps = Pick<MapParameters, 'accessType' | 'groupPermissions'> &
   TemplateMappings & { authResourceName: string };
 
+/**
+ * Class to generate cfn for Map resource
+ */
 export class MapStack extends BaseStack {
   protected readonly groupPermissions: string[];
   protected readonly accessType: string;
@@ -30,7 +33,7 @@ export class MapStack extends BaseStack {
     this.mapRegion = this.regionMapping.findInMap(cdk.Fn.ref('AWS::Region'), 'locationServiceRegion');
 
     const inputParameters: string[] = (this.props.groupPermissions || []).map(
-      (group: string) => `authuserPoolGroups${group}GroupRole`
+      (group: string) => `authuserPoolGroups${group}GroupRole`,
     );
     inputParameters.push(
       `auth${this.authResourceName}UserPoolId`,
@@ -39,7 +42,7 @@ export class MapStack extends BaseStack {
       'mapName',
       'mapStyle',
       'env',
-      'isDefault'
+      'isDefault',
     );
     this.parameters = this.constructInputParameters(inputParameters);
 
@@ -49,16 +52,20 @@ export class MapStack extends BaseStack {
     this.constructOutputs();
   }
 
-  private constructOutputs() {
+  private constructOutputs(): void {
+    // eslint-disable-next-line no-new
     new cdk.CfnOutput(this, 'Name', {
       value: this.mapResource.getAtt('MapName').toString(),
     });
+    // eslint-disable-next-line no-new
     new cdk.CfnOutput(this, 'Style', {
       value: this.parameters.get('mapStyle')!.valueAsString,
     });
+    // eslint-disable-next-line no-new
     new cdk.CfnOutput(this, 'Region', {
       value: this.mapRegion,
     });
+    // eslint-disable-next-line no-new
     new cdk.CfnOutput(this, 'Arn', {
       value: this.mapResource.getAtt('MapArn').toString(),
     });
@@ -100,7 +107,7 @@ export class MapStack extends BaseStack {
       resourceType: 'Custom::LambdaCallout',
       properties: {
         mapName: this.mapName,
-        mapStyle: mapStyle,
+        mapStyle,
         region: this.mapRegion,
         env: cdk.Fn.ref('env'),
       },
@@ -111,32 +118,33 @@ export class MapStack extends BaseStack {
 
   // Grant read-only access to the Map for Authorized and/or Guest users
   private constructMapPolicyResource(mapResource: cdk.CustomResource): CfnResource {
-    let policy = new iam.PolicyDocument({
+    const policy = new iam.PolicyDocument({
       statements: [
         new iam.PolicyStatement({
           effect: iam.Effect.ALLOW,
+          // eslint-disable-next-line spellcheck/spell-checker
           actions: ['geo:GetMapStyleDescriptor', 'geo:GetMapGlyphs', 'geo:GetMapSprites', 'geo:GetMapTile'],
           resources: [mapResource.getAtt('MapArn').toString()],
         }),
       ],
     });
 
-    let cognitoRoles: Array<string> = new Array();
-    if (this.accessType === AccessType.AuthorizedUsers ||
-      this.accessType === AccessType.AuthorizedAndGuestUsers) {
+    const cognitoRoles: Array<string> = [];
+    if (this.accessType === AccessType.AuthorizedUsers
+      || this.accessType === AccessType.AuthorizedAndGuestUsers) {
       cognitoRoles.push(this.parameters.get('authRoleName')!.valueAsString);
     }
-    if (this.accessType == AccessType.AuthorizedAndGuestUsers) {
+    if (this.accessType === AccessType.AuthorizedAndGuestUsers) {
       cognitoRoles.push(this.parameters.get('unauthRoleName')!.valueAsString);
     }
     if (this.groupPermissions && this.authResourceName) {
       this.groupPermissions.forEach((group: string) => {
         cognitoRoles.push(
           cdk.Fn.join('-',
-          [
+            [
             this.parameters.get(`auth${this.authResourceName}UserPoolId`)!.valueAsString,
-            `${group}GroupRole`
-          ])
+            `${group}GroupRole`,
+            ]),
         );
       });
     }