+++ packages/amplify-e2e-core/src/categories/auth.ts
@@ -1,11 +1,19 @@
 import _ from 'lodash';
-import { getCLIPath, getSocialProviders, KEY_DOWN_ARROW, KEY_UP_ARROW, nspawn as spawn, setTransformerVersionFlag } from '..';
+import {
+  getCLIPath, getSocialProviders, KEY_DOWN_ARROW, KEY_UP_ARROW, nspawn as spawn, setTransformerVersionFlag,
+} from '..';
 
+/**
+ *
+ */
 export type AddAuthUserPoolOnlyNoOAuthSettings = {
   resourceName: string;
   userPoolName: string;
 };
 
+/**
+ *
+ */
 export type AddAuthUserPoolOnlyWithOAuthSettings = AddAuthUserPoolOnlyNoOAuthSettings & {
   domainPrefix: string;
   signInUrl1: string;
@@ -24,6 +32,9 @@ export type AddAuthUserPoolOnlyWithOAuthSettings = AddAuthUserPoolOnlyNoOAuthSet
   appleAppPrivateKey: string;
 };
 
+/**
+ *
+ */
 export type AddAuthIdentityPoolAndUserPoolWithOAuthSettings = AddAuthUserPoolOnlyWithOAuthSettings & {
   identityPoolName: string;
   allowUnauthenticatedIdentities: boolean;
@@ -34,9 +45,12 @@ export type AddAuthIdentityPoolAndUserPoolWithOAuthSettings = AddAuthUserPoolOnl
   idpAppleAppId: string;
 };
 
-export function addAuthWithDefault(cwd: string, settings: any = {}): Promise<void> {
+/**
+ *
+ */
+export function addAuthWithDefault(cwd: string, settings: any = {}, testingWithLatestCodebase = false): Promise<void> {
   return new Promise((resolve, reject) => {
-    spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
+    spawn(getCLIPath(testingWithLatestCodebase), ['add', 'auth'], { cwd, stripColors: true })
       .wait('Do you want to use the default authentication')
       .sendCarriageReturn()
       .wait('How do you want users to be able to sign in')
@@ -54,6 +68,9 @@ export function addAuthWithDefault(cwd: string, settings: any = {}): Promise<voi
   });
 }
 
+/**
+ *
+ */
 export function runAmplifyAuthConsole(cwd: string): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['auth', 'console'], { cwd, stripColors: true })
@@ -70,6 +87,9 @@ export function runAmplifyAuthConsole(cwd: string): Promise<void> {
   });
 }
 
+/**
+ *
+ */
 export function removeAuthWithDefault(cwd: string): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['remove', 'auth'], { cwd, stripColors: true })
@@ -88,6 +108,9 @@ export function removeAuthWithDefault(cwd: string): Promise<void> {
   });
 }
 
+/**
+ *
+ */
 export function addAuthWithGroupTrigger(cwd: string, settings: any): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
@@ -131,6 +154,9 @@ const defaultOptions: AddApiOptions = {
   transformerVersion: 2,
 };
 
+/**
+ *
+ */
 export function addAuthViaAPIWithTrigger(cwd: string, opts: Partial<AddApiOptions> = {}): Promise<void> {
   const options = _.assign(defaultOptions, opts);
   return new Promise((resolve, reject) => {
@@ -181,6 +207,9 @@ export function addAuthViaAPIWithTrigger(cwd: string, opts: Partial<AddApiOption
   });
 }
 
+/**
+ *
+ */
 export function addAuthwithUserPoolGroupsViaAPIWithTrigger(cwd: string, opts: Partial<AddApiOptions> = {}): Promise<void> {
   const options = _.assign(defaultOptions, opts);
   return new Promise((resolve, reject) => {
@@ -270,6 +299,9 @@ export function addAuthwithUserPoolGroupsViaAPIWithTrigger(cwd: string, opts: Pa
   });
 }
 
+/**
+ *
+ */
 export function addAuthWithCustomTrigger(cwd: string, settings: any): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
@@ -350,6 +382,9 @@ export function addAuthWithCustomTrigger(cwd: string, settings: any): Promise<vo
   });
 }
 
+/**
+ *
+ */
 export function updateAuthSignInSignOutUrl(cwd: string, settings: any): Promise<void> {
   const testingWithLatestCodebase = settings.testingWithLatestCodebase ?? false;
   return new Promise((resolve, reject) => {
@@ -390,6 +425,9 @@ export function updateAuthSignInSignOutUrl(cwd: string, settings: any): Promise<
   });
 }
 
+/**
+ *
+ */
 export function updateAuthToRemoveFederation(cwd: string, settings: any): Promise<void> {
   const testingWithLatestCodebase = settings.testingWithLatestCodebase ?? false;
   return new Promise((resolve, reject) => {
@@ -412,6 +450,9 @@ export function updateAuthToRemoveFederation(cwd: string, settings: any): Promis
   });
 }
 
+/**
+ *
+ */
 export function updateAuthWithoutCustomTrigger(cwd: string, settings: any): Promise<void> {
   const testingWithLatestCodebase = settings.testingWithLatestCodebase ?? false;
   return new Promise((resolve, reject) => {
@@ -472,6 +513,9 @@ export function updateAuthWithoutCustomTrigger(cwd: string, settings: any): Prom
   });
 }
 
+/**
+ *
+ */
 export function addAuthWithRecaptchaTrigger(cwd: string, settings: any): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
@@ -505,6 +549,9 @@ export function addAuthWithRecaptchaTrigger(cwd: string, settings: any): Promise
   });
 }
 
+/**
+ *
+ */
 export function updateAuthRemoveRecaptchaTrigger(cwd: string, settings: any): Promise<void> {
   const testingWithLatestCodebase = settings.testingWithLatestCodebase ?? false;
   console.log(testingWithLatestCodebase);
@@ -564,6 +611,9 @@ export function updateAuthRemoveRecaptchaTrigger(cwd: string, settings: any): Pr
   });
 }
 
+/**
+ *
+ */
 export function addAuthWithSignInSignOutUrl(cwd: string, settings: any): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
@@ -600,10 +650,14 @@ export function addAuthWithSignInSignOutUrl(cwd: string, settings: any): Promise
   });
 }
 
+/**
+ *
+ */
 export function addAuthWithDefaultSocial_v4_30(cwd: string, settings: any): Promise<void> {
   return new Promise((resolve, reject) => {
-    const { FACEBOOK_APP_ID, FACEBOOK_APP_SECRET, GOOGLE_APP_ID, GOOGLE_APP_SECRET, AMAZON_APP_ID, AMAZON_APP_SECRET } =
-      getSocialProviders(true);
+    const {
+      FACEBOOK_APP_ID, FACEBOOK_APP_SECRET, GOOGLE_APP_ID, GOOGLE_APP_SECRET, AMAZON_APP_ID, AMAZON_APP_SECRET,
+    } = getSocialProviders(true);
 
     spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
       .wait('Do you want to use the default authentication and security configuration?')
@@ -655,6 +709,9 @@ export function addAuthWithDefaultSocial_v4_30(cwd: string, settings: any): Prom
   });
 }
 
+/**
+ *
+ */
 export function addAuthWithDefaultSocial(cwd: string, settings: any): Promise<void> {
   return new Promise((resolve, reject) => {
     const {
@@ -732,6 +789,9 @@ export function addAuthWithDefaultSocial(cwd: string, settings: any): Promise<vo
   });
 }
 
+/**
+ *
+ */
 export function addAuthUserPoolOnly(cwd: string, settings: any): Promise<void> {
   return new Promise((resolve, reject) => {
     const {
@@ -893,6 +953,9 @@ export function addAuthUserPoolOnly(cwd: string, settings: any): Promise<void> {
 }
 
 // creates 2 groups: Admins, Users
+/**
+ *
+ */
 export function addAuthWithGroups(cwd: string): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
@@ -959,6 +1022,9 @@ export function addAuthWithGroups(cwd: string): Promise<void> {
 }
 
 // creates 2 groups: Admins, Users
+/**
+ *
+ */
 export function addAuthWithGroupsAndAdminAPI(cwd: string, settings?: any): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
@@ -1033,6 +1099,9 @@ export function addAuthWithGroupsAndAdminAPI(cwd: string, settings?: any): Promi
   });
 }
 
+/**
+ *
+ */
 export function addAuthWithMaxOptions(cwd: string, settings: any): Promise<void> {
   const {
     FACEBOOK_APP_ID,
@@ -1238,7 +1307,10 @@ export function addAuthWithMaxOptions(cwd: string, settings: any): Promise<void>
   });
 }
 
-//add default auth with pre token generation trigger
+// add default auth with pre token generation trigger
+/**
+ *
+ */
 export function addAuthWithPreTokenGenerationTrigger(projectDir: string): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['add', 'auth'], { cwd: projectDir, stripColors: true })
@@ -1251,7 +1323,7 @@ export function addAuthWithPreTokenGenerationTrigger(projectDir: string): Promis
       .wait('What attributes are required for signing up?')
       .sendCarriageReturn()
       .wait('Do you want to enable any of the following capabilities')
-      .send(KEY_UP_ARROW) //Override ID Token Claims
+      .send(KEY_UP_ARROW) // Override ID Token Claims
       .sendLine(' ')
       .wait('Successfully added the Lambda function locally')
       .wait('Do you want to edit your alter-claims function now')
@@ -1266,6 +1338,9 @@ export function addAuthWithPreTokenGenerationTrigger(projectDir: string): Promis
   });
 }
 
+/**
+ *
+ */
 export function updateAuthAddUserGroups(projectDir: string, groupNames: string[], settings?: any): Promise<void> {
   if (groupNames.length == 0) {
     return;
@@ -1314,6 +1389,9 @@ export function updateAuthAddUserGroups(projectDir: string, groupNames: string[]
   });
 }
 
+/**
+ *
+ */
 export function addAuthUserPoolOnlyWithOAuth(cwd: string, settings: AddAuthUserPoolOnlyWithOAuthSettings): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
@@ -1412,12 +1490,15 @@ export function addAuthUserPoolOnlyWithOAuth(cwd: string, settings: AddAuthUserP
   });
 }
 
+/**
+ *
+ */
 export function addAuthIdentityPoolAndUserPoolWithOAuth(
   cwd: string,
   settings: AddAuthIdentityPoolAndUserPoolWithOAuthSettings,
 ): Promise<void> {
   return new Promise((resolve, reject) => {
-    let chain = spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
+    const chain = spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
       .wait('Do you want to use the default authentication and security configuration?')
       .sendKeyDown(2)
       .sendCarriageReturn()
@@ -1536,6 +1617,9 @@ export function addAuthIdentityPoolAndUserPoolWithOAuth(
   });
 }
 
+/**
+ *
+ */
 export function addAuthUserPoolOnlyNoOAuth(cwd: string, settings: AddAuthUserPoolOnlyNoOAuthSettings): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['add', 'auth'], { cwd, stripColors: true })
@@ -1591,7 +1675,10 @@ export function addAuthUserPoolOnlyNoOAuth(cwd: string, settings: AddAuthUserPoo
   });
 }
 
-export function updateAuthAddAdminQueries(projectDir: string, groupName: string = 'adminQueriesGroup', settings: any = {}): Promise<void> {
+/**
+ *
+ */
+export function updateAuthAddAdminQueries(projectDir: string, groupName = 'adminQueriesGroup', settings: any = {}): Promise<void> {
   const testingWithLatestCodebase = settings.testingWithLatestCodebase ?? false;
   return new Promise((resolve, reject) => {
     const chain = spawn(getCLIPath(testingWithLatestCodebase), ['update', 'auth'], { cwd: projectDir, stripColors: true });
@@ -1620,6 +1707,9 @@ export function updateAuthAddAdminQueries(projectDir: string, groupName: string 
   });
 }
 
+/**
+ *
+ */
 export function updateAuthWithoutTrigger(cwd: string, settings: any): Promise<void> {
   const testingWithLatestCodebase = settings.testingWithLatestCodebase ?? false;
   return new Promise((resolve, reject) => {
@@ -1671,6 +1761,9 @@ export function updateAuthWithoutTrigger(cwd: string, settings: any): Promise<vo
   });
 }
 
+/**
+ *
+ */
 export function updateAuthAdminQueriesWithExtMigration(cwd: string, settings: { testingWithLatestCodebase: boolean }): Promise<void> {
   return spawn(getCLIPath(settings.testingWithLatestCodebase), ['update', 'auth'], { cwd, stripColors: true })
     .wait('Do you want to migrate auth resource')
@@ -1690,6 +1783,9 @@ export function updateAuthAdminQueriesWithExtMigration(cwd: string, settings: { 
     .runAsync();
 }
 
+/**
+ *
+ */
 export function updateAuthMFAConfiguration(projectDir: string, settings: any = {}): Promise<void> {
   return spawn(getCLIPath(settings.testingWithLatestCodebase), ['update', 'auth'], { cwd: projectDir, stripColors: true })
     .wait('What do you want to do?')