+++ packages/amplify-provider-awscloudformation/src/root-stack-builder/root-stack-transform.ts
@@ -10,6 +10,9 @@ import * as vm from 'vm2';
 import { AmplifyRootStack, AmplifyRootStackOutputs } from './root-stack-builder';
 import { RootStackSynthesizer } from './stack-synthesizer';
 
+/**
+ * class to manage stack lifecycle
+ */
 export class AmplifyRootStackTransform {
   private app: cdk.App | undefined;
   private _rootTemplateObj: AmplifyRootStack; // Props to modify Root stack data
@@ -180,8 +183,7 @@ export class AmplifyRootStackTransform {
   };
 
   /**
-   *
-   * @returns return CFN templates synthesized by app
+   * return CFN templates synthesized by app
    */
   private synthesizeTemplates = async (): Promise<Template> => {
     this.app?.synth();
@@ -203,6 +205,9 @@ export class AmplifyRootStackTransform {
     });
   };
 
+  /**
+   * return root stack
+   */
   public getRootStack(): AmplifyRootStack {
     if (this._rootTemplateObj) {
       return this._rootTemplateObj;