+++ packages/amplify-provider-awscloudformation/src/root-stack-builder/stack-synthesizer.ts
@@ -25,16 +25,14 @@ export class RootStackSynthesizer extends LegacyStackSynthesizer {
   /**
    * set a specific stack
    */
-  setStackAsset(templateName: string, template: string): void {
+  setStackAsset = (templateName: string, template: string): void => {
     RootStackSynthesizer.stackAssets.set(templateName, JSONUtilities.parse(template));
   }
 
   /**
    * get all stacks
    */
-  collectStacks(): Map<string, Template> {
-    return new Map(RootStackSynthesizer.stackAssets.entries());
-  }
+  collectStacks = (): Map<string, Template> => new Map(RootStackSynthesizer.stackAssets.entries())
 
   /**
    * add a stack