+++ packages/amplify-util-uibuilder/src/commands/utils/codegenResources.ts
@@ -10,14 +10,19 @@ import {
   RenderTextComponentResponse,
   GenericDataSchema,
   getGenericFromDataStore,
+  StudioForm,
+  StudioSchema,
 } from '@aws-amplify/codegen-ui';
 import {
   AmplifyRenderer,
+  AmplifyFormRenderer,
   ReactThemeStudioTemplateRenderer,
   ReactIndexStudioTemplateRenderer,
   ModuleKind,
   ScriptTarget,
   ScriptKind,
+  UtilTemplateType,
+  ReactUtilsStudioTemplateRenderer,
 } from '@aws-amplify/codegen-ui-react';
 import { printer } from 'amplify-prompts';
 import {
@@ -84,14 +89,14 @@ export const createUiBuilderTheme = (context: $TSContext, schema: StudioTheme): 
 };
 
 /**
- * Writes index file to the work space
+ * Writes form file to the work space
  */
-export const generateAmplifyUiBuilderIndexFile = (context: $TSContext, schemas: StudioComponent[]): RenderTextComponentResponse => {
+export const createUiBuilderForm = (context: $TSContext, schema: StudioForm, dataSchema?: GenericDataSchema): StudioForm => {
   const uiBuilderComponentsPath = getUiBuilderComponentsPath(context);
   const rendererFactory = new StudioTemplateRendererFactory(
-    (component: StudioComponent[]) => (new ReactIndexStudioTemplateRenderer(component, config) as unknown) as StudioTemplateRenderer<
+    (form: StudioForm) => (new AmplifyFormRenderer(form, dataSchema, config) as unknown) as StudioTemplateRenderer<
         unknown,
-        StudioComponent[],
+        StudioForm,
         FrameworkOutputManager<unknown>,
         RenderTextComponentResponse
       >,
@@ -104,7 +109,81 @@ export const generateAmplifyUiBuilderIndexFile = (context: $TSContext, schemas: 
   });
 
   try {
-    return rendererManager.renderSchemaToTemplate(schemas);
+    rendererManager.renderSchemaToTemplate(schema);
+    return schema;
+  } catch (e) {
+    printer.debug(e);
+    printer.debug('Skipping invalid form with schema');
+    printer.debug(JSON.stringify(schema, null, 2));
+    throw e;
+  }
+};
+
+/**
+ * Writes index file to the work space
+ */
+export const generateAmplifyUiBuilderIndexFile = (context: $TSContext, schemas: StudioSchema[]): void => {
+  const uiBuilderComponentsPath = getUiBuilderComponentsPath(context);
+  const rendererFactory = new StudioTemplateRendererFactory(
+    (schema: StudioSchema[]) => (new ReactIndexStudioTemplateRenderer(schema, config) as unknown) as StudioTemplateRenderer<
+        unknown,
+        StudioSchema[],
+        FrameworkOutputManager<unknown>,
+        RenderTextComponentResponse
+      >,
+  );
+
+  const outputPathDir = uiBuilderComponentsPath;
+
+  const rendererManager = new StudioTemplateRendererManager(rendererFactory, {
+    outputPathDir,
+  });
+
+  try {
+    rendererManager.renderSchemaToTemplate(schemas);
+  } catch (e) {
+    printer.debug(e);
+    printer.debug('Failed to generate component index file');
+    throw e;
+  }
+};
+
+type UtilFileChecks = {
+  hasForms: boolean;
+  hasViews: boolean;
+};
+
+/**
+ * Writes utils file to the work space
+ */
+export const generateAmplifyUiBuilderUtilFile = (context: $TSContext, { hasForms, hasViews }: UtilFileChecks): void => {
+  const uiBuilderComponentsPath = getUiBuilderComponentsPath(context);
+  const rendererFactory = new StudioTemplateRendererFactory(
+    (utils: UtilTemplateType[]) => (new ReactUtilsStudioTemplateRenderer(utils, config)),
+  );
+
+  const outputPathDir = uiBuilderComponentsPath;
+
+  const rendererManager = new StudioTemplateRendererManager(rendererFactory, {
+    outputPathDir,
+  });
+
+  const utils = new Set<UtilTemplateType>();
+
+  if (hasForms) {
+    utils.add('validation');
+    utils.add('formatter');
+    utils.add('fetchByPath');
+  }
+
+  if (hasViews) {
+    utils.add('formatter');
+  }
+
+  try {
+    if (utils.size) {
+      rendererManager.renderSchemaToTemplate([...utils]);
+    }
   } catch (e) {
     printer.debug(e);
     printer.debug('Failed to generate component index file');
@@ -119,6 +198,9 @@ export const generateAmplifyUiBuilderIndexFile = (context: $TSContext, schemas: 
 export const getAmplifyDataSchema = async (
   studioClient: AmplifyStudioClient,
 ): Promise<GenericDataSchema | undefined> => {
+  if (!studioClient.isGraphQLSupported) {
+    return undefined;
+  }
   try {
     const meta = stateManager.getMeta();
     const resourceName = Object.entries(meta[AmplifyCategories.API]).find(
@@ -137,4 +219,26 @@ export const getAmplifyDataSchema = async (
     printer.debug(error.toString());
     return undefined;
   }
+};
+
+/**
+ * generates base create/update froms from names
+ */
+export const generateBaseForms = (modelMap: {[model: string]: Set<'create' | 'update'>}): StudioForm[] => {
+  const getSchema = (name: string, type: 'create' | 'update') : StudioForm => ({
+    name: `${name}${type === 'create' ? 'CreateForm' : 'UpdateForm'}`,
+    formActionType: type,
+    dataType: { dataSourceType: 'DataStore', dataTypeName: name },
+    fields: {},
+    sectionalElements: {},
+    style: {},
+    cta: {},
+  });
+
+  const schemas : StudioForm[] = [];
+
+  Object.entries(modelMap).forEach(([name, set]) => {
+    set.forEach(type => schemas.push(getSchema(name, type)));
+  });
+  return schemas;
 };